+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/forging/ForgeResultSlot.java
@@ -1,6 +1,5 @@
 package insane96mcp.iguanatweaksexpanded.module.mining.forging;
 
-import insane96mcp.iguanatweaksexpanded.module.mining.multiblockfurnaces.block.ForgeBlockEntity;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;